requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
replicate==0.27.0
